scraper_refactor_pre_kill
