+++ kifi-backend/modules/scraper/app/com/keepit/scraper/ScrapeProcessor.scala
@@ -1,7 +1,7 @@
 package com.keepit.scraper
 
 import com.google.inject._
-import play.api.{Play, Plugin}
+import play.api.Play
 import com.keepit.common.logging.Logging
 import com.keepit.common.akka.FortyTwoActor
 import com.keepit.common.healthcheck.AirbrakeNotifier
@@ -9,51 +9,47 @@ import com.keepit.model._
 import com.keepit.scraper.extractor._
 import com.keepit.search.{LangDetector, Article, ArticleStore}
 import com.keepit.common.store.S3ScreenshotStore
-import com.keepit.shoebox.ShoeboxServiceClient
 import java.io.File
-import scala.concurrent.{Awaitable, Future, Await}
+import scala.concurrent.Future
 import scala.concurrent.duration._
 import org.joda.time.Days
 import com.keepit.common.time._
 import com.keepit.common.net.URI
 import org.apache.http.HttpStatus
 import com.keepit.scraper.mediatypes.MediaTypes
-import com.keepit.common.db.Id
 import akka.actor._
 import akka.pattern.ask
 import akka.util.Timeout
-import play.api.libs.concurrent.Execution.Implicits.defaultContext
-import akka.routing.{SmallestMailboxRouter}
+import akka.routing.SmallestMailboxRouter
 import play.api.Play.current
-import scala.Some
 import scala.util.Success
 
 
-@ProvidedBy(classOf[ScrapeProcessorPluginProvider])
-trait ScrapeProcessorPlugin extends Plugin {
+@ProvidedBy(classOf[ScrapeProcessorProvider])
+trait ScrapeProcessor {
   def fetchBasicArticle(url:String, proxyOpt:Option[HttpProxy], extractorProviderTypeOpt:Option[ExtractorProviderType]):Future[Option[BasicArticle]]
   def scrapeArticle(uri:NormalizedURI, info:ScrapeInfo, proxyOpt:Option[HttpProxy]):Future[(NormalizedURI, Option[Article])]
   def asyncScrape(uri:NormalizedURI, info:ScrapeInfo, proxyOpt:Option[HttpProxy]): Unit
 }
 
 @Singleton
-class ScrapeProcessorPluginProvider @Inject() (
+class ScrapeProcessorProvider @Inject() (
   scraperConfig: ScraperConfig,
-  asyncScrapeProcessor:AsyncScrapeProcessorPlugin,
-  syncScrapeProcessor:ScrapeProcessorPluginImpl
-) extends Provider[ScrapeProcessorPlugin] with Logging {
+  asyncScrapeProcessor:AsyncScrapeProcessor,
+  syncScrapeProcessor:SyncScrapeProcessor
+) extends Provider[ScrapeProcessor] with Logging {
 
   lazy val plugin = if (scraperConfig.async) asyncScrapeProcessor else syncScrapeProcessor
-  log.info(s"[ScrapeProcessorPluginProvider] created with: $scraperConfig")
+  log.info(s"[ScrapeProcessorProvider] created with: $scraperConfig")
 
   def get = plugin
 }
 
-class ScrapeProcessorPluginImpl @Inject() (
+class SyncScrapeProcessor @Inject() (
   scraperConfig: ScraperConfig,
   sysProvider:   Provider[ActorSystem],
-  procProvider:  Provider[ScrapeProcessor]
-) extends ScrapeProcessorPlugin with Logging {
+  procProvider:  Provider[SyncScraperActor]
+) extends ScrapeProcessor with Logging {
 
   lazy val system = sysProvider.get
   lazy val actor = {
@@ -80,7 +76,7 @@ case class AsyncScrape(uri:NormalizedURI, info:ScrapeInfo, proxyOpt:Option[HttpP
 case class FetchBasicArticle(url:String, proxyOpt:Option[HttpProxy], extractorProviderTypeOpt:Option[ExtractorProviderType])
 case class ScrapeArticle(uri:NormalizedURI, info:ScrapeInfo, proxyOpt:Option[HttpProxy])
 
-class ScrapeProcessor @Inject() (
+class SyncScraperActor @Inject() (
   airbrake:AirbrakeNotifier,
   config: ScraperConfig,
   syncScraper: SyncScraper,
@@ -383,56 +379,3 @@ class SyncScraper @Inject() (
 
 }
 
-class ShoeboxDbCallbackHelper @Inject() (config:ScraperConfig, shoeboxServiceClient:ShoeboxServiceClient) extends SyncShoeboxDbCallbacks with ShoeboxDbCallbacks {
-  implicit val serviceCallTimeout = config.serviceCallTimeout
-
-  private def await[T](awaitable: Awaitable[T]) = Await.result(awaitable, config.syncAwaitTimeout seconds)
-
-  def syncIsUnscrapableP(url: String, destinationUrl: Option[String]) = await(isUnscrapableP(url, destinationUrl))
-  def syncGetNormalizedUri(uri:NormalizedURI):Option[NormalizedURI] = await(getNormalizedUri(uri))
-  def syncSaveNormalizedUri(uri:NormalizedURI):NormalizedURI = await(saveNormalizedUri(uri))
-  def syncSaveScrapeInfo(info:ScrapeInfo):ScrapeInfo = await(saveScrapeInfo(info))
-  def syncGetBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]):Seq[Bookmark] = await(getBookmarksByUriWithoutTitle(uriId))
-  def syncGetLatestBookmark(uriId: Id[NormalizedURI]): Option[Bookmark] = await(getLatestBookmark(uriId))
-  def syncRecordPermanentRedirect(uri: NormalizedURI, redirect: HttpRedirect): NormalizedURI = await(recordPermanentRedirect(uri, redirect))
-  def syncSaveBookmark(bookmark:Bookmark):Bookmark = await(saveBookmark(bookmark))
-
-  def getNormalizedUri(uri:NormalizedURI):Future[Option[NormalizedURI]] = {
-    uri.id match {
-      case Some(id) => shoeboxServiceClient.getNormalizedURI(id).map(Some(_))
-      case None => shoeboxServiceClient.getNormalizedURIByURL(uri.url)
-    }
-  }
-  def saveNormalizedUri(uri:NormalizedURI):Future[NormalizedURI] = shoeboxServiceClient.saveNormalizedURI(uri)
-  def saveScrapeInfo(info:ScrapeInfo):Future[ScrapeInfo] = shoeboxServiceClient.saveScrapeInfo(if (info.state == ScrapeInfoStates.INACTIVE) info else info.withState(ScrapeInfoStates.ACTIVE))
-  def getBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]):Future[Seq[Bookmark]] = shoeboxServiceClient.getBookmarksByUriWithoutTitle(uriId)
-  def getLatestBookmark(uriId: Id[NormalizedURI]): Future[Option[Bookmark]] = shoeboxServiceClient.getLatestBookmark(uriId)
-  def saveBookmark(bookmark:Bookmark): Future[Bookmark] = shoeboxServiceClient.saveBookmark(bookmark)
-  def recordPermanentRedirect(uri: NormalizedURI, redirect: HttpRedirect): Future[NormalizedURI] = shoeboxServiceClient.recordPermanentRedirect(uri, redirect)
-  def isUnscrapableP(url: String, destinationUrl: Option[String]) = shoeboxServiceClient.isUnscrapableP(url, destinationUrl)
-
-}
-
-@ImplementedBy(classOf[ShoeboxDbCallbackHelper])
-trait SyncShoeboxDbCallbacks {
-  def syncIsUnscrapableP(url: String, destinationUrl: Option[String]):Boolean
-  def syncGetNormalizedUri(uri:NormalizedURI):Option[NormalizedURI]
-  def syncSaveNormalizedUri(uri:NormalizedURI):NormalizedURI
-  def syncSaveScrapeInfo(info:ScrapeInfo):ScrapeInfo
-  def syncGetBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]):Seq[Bookmark]
-  def syncGetLatestBookmark(uriId: Id[NormalizedURI]): Option[Bookmark]
-  def syncRecordPermanentRedirect(uri: NormalizedURI, redirect: HttpRedirect): NormalizedURI
-  def syncSaveBookmark(bookmark:Bookmark):Bookmark
-}
-
-@ImplementedBy(classOf[ShoeboxDbCallbackHelper])
-trait ShoeboxDbCallbacks {
-  def getNormalizedUri(uri:NormalizedURI):Future[Option[NormalizedURI]]
-  def saveNormalizedUri(uri:NormalizedURI):Future[NormalizedURI]
-  def saveScrapeInfo(info:ScrapeInfo):Future[ScrapeInfo]
-  def getBookmarksByUriWithoutTitle(uriId: Id[NormalizedURI]):Future[Seq[Bookmark]]
-  def getLatestBookmark(uriId: Id[NormalizedURI]): Future[Option[Bookmark]]
-  def saveBookmark(bookmark:Bookmark): Future[Bookmark]
-  def recordPermanentRedirect(uri: NormalizedURI, redirect: HttpRedirect): Future[NormalizedURI]
-  def isUnscrapableP(url: String, destinationUrl: Option[String]): Future[Boolean]
-}
