+++ kifi-backend/modules/scraper/test/com/keepit/scraper/ScraperModuleTest.scala
@@ -21,7 +21,7 @@ import java.io.File
 case class ScraperTestModule() extends ScraperServiceModule (
   cacheModule = ScraperCacheModule(HashMapMemoryCacheModule()),
   storeModule = ScraperTestStoreModule(),
-  scraperModule = ScraperImplModule()
+  scrapeProcessorModule = ScrapeProcessorImplModule()
 ) with CommonDevModule { }
